Massive reorganizing, avalon() is now properly being exported to server.js
