--- cloudStorage.js
@@ -1,3 +1,0 @@
-
-
-modules.exports  { createAvalonian }
