+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "discord-1337-bot",
   "version": "0.0.1",
-  "main": "server.js",
+  "main": "src/server.js",
   "scripts": {
     "start": "node server.js"
   },
