--- tools/tools.js
@@ -1,5 +1,0 @@
-module.exports = function randomNumberInRange(range) {
-  return Math.floor(Math.random() * Math.floor(range));
-}
-
-exports.randomNumberInRange = this.randomNumberInRange;
